refactor: common 패키지 정리
- 특정 Domain이 아닌 공통으로 사용되는 모듈 common 패키지로 이동
