+++ wecare-backend/wecare/src/test/java/com/example/wecare/unit/service/InvitationServiceTest.java
@@ -7,7 +7,7 @@ import com.example.wecare.invitation.service.InvitationService;
 import com.example.wecare.member.domain.Member;
 import com.example.wecare.member.domain.Role;
 import com.example.wecare.member.repository.MemberRepository;
-import com.example.wecare.redis.service.RedisService;
+import com.example.wecare.common.service.RedisService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
